+++ yorm/bases/__init__.py
@@ -1,6 +1,4 @@
 """Base classes for mapping and conversion."""
 
-from .container import Container
-from .converter import Converter
 from .mappable import Mappable
-from .convertible import Convertible
+from .converter import Converter, Container